WP-9 Implemented - package, class name refactored - min sdk version set to 19 - add action bar to mainactivity
