+++ app/build.gradle
@@ -47,7 +47,7 @@ android {
 
     defaultConfig {
         applicationId "com.pachalenlabs.wallp"
-        minSdkVersion 22
+        minSdkVersion 19
         targetSdkVersion 23
         versionCode 1
         versionName "1.0"
