--- app/src/main/java/com/pachalenlabs/wallp/ActMain.java
@@ -1,11 +1,0 @@
-package com.pachalenlabs.wallp;
-
-import android.app.Activity;
-
-import org.androidannotations.annotations.EActivity;
-
-@EActivity(R.layout.activity_act_main)
-public class ActMain extends Activity {
-
-    
-}
