+++ app/src/main/java/com/pachalenlabs/wallp/WPApp.java
@@ -1,4 +1,4 @@
-package com.pachalenlabs.wallp.module;
+package com.pachalenlabs.wallp;
 
 import android.app.Application;
