+++ app/src/main/java/com/pachalenlabs/wallp/ui/activity/MainActivity.java
@@ -1,0 +1,13 @@
+package com.pachalenlabs.wallp.ui.activity;
+
+import android.support.v7.app.AppCompatActivity;
+
+import com.pachalenlabs.wallp.R;
+
+import org.androidannotations.annotations.EActivity;
+
+@EActivity(R.layout.activity_act_main)
+public class MainActivity extends AppCompatActivity {
+
+    
+}
