+++ app/src/main/res/layout/activity_act_logo.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ActLogo"
+    tools:context=".ui.activity.LogoActivity"
     android:background="#bd6969">
 
     <android.support.design.widget.AppBarLayout
